Update README.md
The README.md file has been updated to include the language in each code block. This makes it easier for users to understand the code and increases the readability of the document.

Examples of code blocks have been updated from ``` to ```Kotlin to indicate that the code is written in Kotlin.